pyproject.toml
[project]
name = "team-corgi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.7.9",
    "click==8.2.1",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi-cli==0.0.8",
    "fastapi-cloud-cli==0.1.2",
    "fastapi[standard]==0.115.4",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "psycopg2==2.9.10",
    "pydantic==2.8.2",
    "pydantic-core==2.20.1",
    "pygments==2.19.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich==14.0.0",
    "rich-toolkit==0.14.8",
    "rignore==0.5.1",
    "sentry-sdk==2.32.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "sqlmodel==0.0.24",
    "starlette==0.41.3",
    "typer==0.16.0",
    "typing-extensions==4.14.1",
    "urllib3==2.5.0",
    "uvicorn==0.30.6",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
]
